+++ src/elements/mod.rs
@@ -22,7 +22,6 @@ pub mod force_deployment;
 pub mod governance_stage1_calls;
 pub mod governance_stage2_calls;
 pub mod initialize_data_new_chain;
-pub mod post_upgrade_calldata;
 pub mod protocol_version;
 pub mod set_new_version_upgrade;
 
@@ -33,17 +32,25 @@ pub struct UpgradeOutput {
     pub(crate) create2_factory_salt: FixedBytes<32>,
     pub(crate) deployer_addr: Address,
     pub(crate) era_chain_id: u64,
-    pub(crate) governance_stage1_calls: String,
-    pub(crate) governance_stage2_calls: String,
+
+    pub(crate) governance_calls: GovernanceCalls,
+
     pub(crate) l1_chain_id: u64,
 
     pub(crate) protocol_upgrade_handler_proxy_address: Address,
     pub(crate) protocol_upgrade_handler_impl_address: Address,
 
+    #[serde(rename = "contracts_newConfig")]
     pub(crate) contracts_config: ContractsConfig,
     pub(crate) deployed_addresses: DeployedAddresses,
 
     pub(crate) transactions: Vec<String>,
+}
+
+#[derive(Debug, Deserialize)]
+pub struct GovernanceCalls {
+    pub(crate) governance_stage1_calls: String,
+    pub(crate) governance_stage2_calls: String,
 }
 
 #[derive(Debug, Deserialize)]
@@ -219,25 +226,34 @@ impl UpgradeOutput {
             .await;
 
         let stage1 = GovernanceStage1Calls {
-            calls: CallList::parse(&self.governance_stage1_calls),
+            calls: CallList::parse(&self.governance_calls.governance_stage1_calls),
         };
+
+        let expected_upgrade_facets = facets_to_remove.merge(facets_to_add.clone()).clone();
 
         stage1
             .verify(
                 &self.deployed_addresses,
                 verifiers,
                 result,
-                facets_to_remove.merge(facets_to_add.clone()),
+                expected_upgrade_facets.clone(),
                 &self.chain_upgrade_diamond_cut,
             )
             .await
             .context("stage1")?;
 
         let stage2 = GovernanceStage2Calls {
-            calls: CallList::parse(&self.governance_stage2_calls),
+            calls: CallList::parse(&self.governance_calls.governance_stage2_calls),
         };
         let (expected_chain_creation_data, expected_force_deployments) = stage2
-            .verify(verifiers, result, facets_to_add)
+            .verify(
+                verifiers,
+                result,
+                facets_to_add,
+                &self.deployed_addresses,
+                expected_upgrade_facets,
+                &self.chain_upgrade_diamond_cut,
+            )
             .await
             .context("stage2")?;