+++ src/main.rs
@@ -93,11 +93,11 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
     if args.display_upgrade_data.unwrap_or_default() {
         println!(
             "Stage1 encoded upgrade data = {}",
-            encode_upgrade_data(&config.governance_stage1_calls)
+            encode_upgrade_data(&config.governance_calls.governance_stage1_calls)
         );
         println!(
             "Stage2 encoded upgrade data = {}",
-            encode_upgrade_data(&config.governance_stage2_calls)
+            encode_upgrade_data(&config.governance_calls.governance_stage2_calls)
         );
     }