+++ CHANGELOG.md
@@ -17,8 +17,8 @@ Added:
 
 **Upgrade**
 
-* stage1 no longer calls `setChainCreationParams`
-* stage2 no longer calls `upgradeAndCall` at the end
+* stage1 is just doing suspendMigrations (and might be fully removed)
+* setChainCreationParams moved to stage2
 
 
 **Other:**
@@ -26,6 +26,11 @@ Added:
 * Bytecode for create2_and_transfer has changed
 * added more 'context' messages - to help with error debugging
 * verify_protocol_upgrade_handler - needs some rework, as it has different behavior on local network.
+
+
+postugprade calldata is now empty (and added force deployment checks into new version tx data check)
+
+